Правки: удалить ненужные комментарии, перенести функцию в отдельный файл
