+++ templates/main.php
@@ -1,14 +1,6 @@
 <?php
-function cut_text (string $str, $length = 300) {
-    if (iconv_strlen($str) <= $length) {
-        return $str;
-    }
-    $str_new = mb_substr($str, 0, $length);
-    $last_whitespace = mb_strripos($str_new, ' ', 0);
-    return mb_substr($str_new, 0, $last_whitespace) . '...';
-};
+require_once('tools.php');
 ?>
-
 <div class="container">
         <h1 class="page__title page__title--popular">Популярное</h1>
     </div>
